# docs/static/assets/openapi.json
{
  "components": {
    "schemas": {
      "DeplyomentTO": {
        "type": "object",
        "required": [
          "target"
        ],
        "properties": {
          "target": {
            "type": "string"
          }
        }
      },
...
      "ArtifactUpdateTO": {
        "type": "object",
        "required": [
          "description",
          "name"
        ],
        "properties": {
          "svgPreview": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          }
        }
      },
      "ArtifactVersionUploadTO": {
        "type": "object",
        "properties": {
          "saveType": {
            "type": "string",
            "enum": [
              "MILESTONE",
              "AUTOSAVE"
            ]
          },
          "xml": {
            "type": "string"
          },
          "versionComment": {
            "type": "string"
          }
        }
      },
...
      "ArtifactTO": {
        "type": "object",
        "required": [
          "description",
          "name",
          "repositoryId"
        ],
        "properties": {
          "createdDate": {
            "format": "date-time",
            "type": "string"
          },
          "svgPreview": {
            "type": "string"
          },
          "repositoryId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string"
          },
          "fileType": {
            "type": "string"
          }
        }
      },
      "AssignmentUpdateTO": {
        "type": "object",
        "required": [
          "repositoryId",
          "role",
          "userId",
          "username"
        ],
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ]
          },
          "repositoryId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "NewArtifactTO": {
        "type": "object",
        "required": [
          "description",
          "name"
        ],
        "properties": {
          "svgPreview": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          }
        }
      },
      "AssignmentTO": {
        "type": "object",
        "required": [
          "repositoryId",
          "role",
          "userId",
          "username"
        ],
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ]
          },
          "repositoryId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "ArtifactVersionTO": {
        "type": "object",
        "required": [
          "artifactId",
          "repositoryId",
          "saveType",
          "updatedDate",
          "xml"
        ],
        "properties": {
          "saveType": {
            "type": "string",
            "enum": [
              "MILESTONE",
              "AUTOSAVE"
            ]
          },
          "milestone": {
            "format": "int32",
            "type": "integer"
          },
          "artifactId": {
            "type": "string"
          },
          "xml": {
            "type": "string"
          },
          "repositoryId": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string"
          }
        }
      }
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
    "/api/assignment/{repositoryId}": {
      "get": {
        "summary": "Get all assigned users",
        "operationId": "getAllAssignedUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentTO"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repositoryId",
            "required": true
          }
        ],
        "tags": [
          "Assignment"
        ]
      }
    },
    "/api/version/{artifactId}/all": {
      "get": {
        "operationId": "getAllVersions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactVersionTO"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Version"
        ]
      }
    },
    "/api/assignment/{repositoryId}/{username}": {
      "delete": {
        "summary": "Delete user assignment",
        "operationId": "deleteUserAssignment",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repositoryId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "username",
            "required": true
          }
        ],
        "tags": [
          "Assignment"
        ]
      }
    },
...
    "/api/artifact/{repositoryId}": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewArtifactTO"
              }
            }
          },
          "required": true
        },
        "operationId": "createArtifact",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactTO"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repositoryId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
...
    "/api/version/{artifactId}/{versionId}/download": {
      "get": {
        "operationId": "downloadVersion",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "versionId",
            "required": true
          }
        ],
        "tags": [
          "Version"
        ]
      }
    },
    "/api/version/{artifactId}/{versionId}": {
      "get": {
        "operationId": "getVersion",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactVersionTO"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "versionId",
            "required": true
          }
        ],
        "tags": [
          "Version"
        ]
      }
    },
...
    "/api/artifact/{artifactId}": {
      "get": {
        "operationId": "getArtifact",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactTO"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      },
      "delete": {
        "summary": "Delete one Artifact and all of its versions",
        "operationId": "deleteArtifact",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactUpdateTO"
              }
            }
          },
          "required": true
        },
        "operationId": "updateArtifact",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactTO"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
...
    "/api/version/{artifactId}": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactVersionUploadTO"
              }
            }
          },
          "required": true
        },
        "operationId": "createOrUpdateVersion",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Version"
        ]
      }
    },
...
    "/api/deploy/{artifactId}/{versionId}": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeplyomentTO"
              }
            }
          },
          "required": true
        },
        "description": "Deploy artifact version",
        "operationId": "deployVersion",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "versionId",
            "required": true
          }
        ],
        "tags": [
          "Deployment"
        ]
      }
    },
    "/api/assignment": {
      "post": {
        "summary": "Create / update user assignment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentUpdateTO"
              }
            }
          },
          "required": true
        },
        "operationId": "createOrUpdateUserAssignment",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Assignment"
        ]
      }
    },
...
    "/api/version/{artifactId}/latest": {
      "get": {
        "summary": "Return the latest version of the requested artifact",
        "operationId": "getLatestVersion",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactVersionTO"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Version"
        ]
      }
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  }
}